+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 [Full changelog](https://github.com/mozilla/glean.js/compare/v0.19.0...main)
 
+* [#696](https://github.com/mozilla/glean.js/pull/696): Expose Node.js entry point `@mozilla/glean/node`.
+* [#695](https://github.com/mozilla/glean.js/pull/695): Implement PlatformInfo module for the Node.js platform.
+
 # v0.19.0 (2021-09-03)
 
 [Full changelog](https://github.com/mozilla/glean.js/compare/v0.18.1...v0.19.0)